Split util.py: move PSIS to diagnostics.py, ECA to new eca.py (7.1)
util.py was 886 lines mixing core utilities, PSIS diagnostics, and
multi-device ECA orchestration. Split into cohesive modules:

- PSIS functions (_gpdfit, _gpinv, psis_weights) → diagnostics.py
- ECA/ensemble functions (eca_step, run_eca, etc.) → new eca.py
- Core utilities remain in util.py (now 458 lines)

Re-exports in util.py preserve backward compatibility. Consumer imports
updated to use canonical locations.
