+++ blackjax/adaptation/laps.py
@@ -25,6 +25,7 @@ from blackjax.adaptation.laps_burn_in import (
     equipartition_diagonal_loss,
 )
 from blackjax.adaptation.step_size import bisection_monotonic_fn
+from blackjax.eca import run_eca
 from blackjax.mcmc.adjusted_mclmc import build_kernel as build_kernel_mclmc
 from blackjax.mcmc.hmc import HMCState
 from blackjax.mcmc.integrators import (
@@ -32,7 +33,6 @@ from blackjax.mcmc.integrators import (
     mclachlan_coefficients,
     omelyan_coefficients,
 )
-from blackjax.util import run_eca
 
 
 class AdaptationState(NamedTuple):
