+++ blackjax/util.py
@@ -3,16 +3,20 @@
 from functools import partial
 from typing import Callable, NamedTuple
 
-import jax
 import jax.numpy as jnp
-from jax import device_put, jit, lax, shard_map, vmap
+from jax import jit, lax
 from jax.flatten_util import ravel_pytree
 from jax.random import normal, split
-from jax.sharding import NamedSharding, PartitionSpec
 from jax.tree_util import tree_leaves, tree_map
 
 from blackjax.base import SamplingAlgorithm, VIAlgorithm
-from blackjax.diagnostics import splitR
+from blackjax.diagnostics import psis_weights as psis_weights  # noqa: F401
+from blackjax.eca import add_all_chains_info as add_all_chains_info  # noqa: F401
+from blackjax.eca import add_splitR as add_splitR  # noqa: F401
+from blackjax.eca import eca_step as eca_step  # noqa: F401
+from blackjax.eca import ensemble_execute_fn as ensemble_execute_fn  # noqa: F401
+from blackjax.eca import run_eca as run_eca  # noqa: F401
+from blackjax.eca import while_with_info as while_with_info  # noqa: F401
 from blackjax.progress_bar import gen_scan_fn
 from blackjax.types import Array, ArrayLikeTree, ArrayTree, PRNGKey
 
@@ -319,300 +323,6 @@ def incremental_value_update(
     return total, average
 
 
-def eca_step(
-    kernel,
-    summary_statistics_fn,
-    adaptation_update,
-    num_chains,
-    superchain_size=None,
-    all_chains_info=None,
-):
-    """
-    Construct a single step of ensemble chain adaptation (eca) to be performed in parallel on multiple devices.
-    """
-
-    def step(state_all, xs):
-        """This function operates on a single device."""
-        (
-            state,
-            adaptation_state,
-        ) = state_all  # state is an array of states, one for each chain on this device. adaptation_state is the same for all chains, so it is not an array.
-        (
-            _,
-            keys_sampling,
-            key_adaptation,
-        ) = xs  # keys_sampling.shape = (chains_per_device, )
-
-        # update the state of all chains on this device
-        state, info = vmap(kernel, (0, 0, None))(keys_sampling, state, adaptation_state)
-
-        # combine all the chains to compute expectation values
-        summary_statistics = vmap(summary_statistics_fn, (0, 0, None))(
-            state, info, key_adaptation
-        )
-        expected_value_summary_statistics = tree_map(
-            lambda summary_statistics: lax.psum(
-                jnp.sum(summary_statistics, axis=0), axis_name="chains"
-            )
-            / num_chains,
-            summary_statistics,
-        )
-
-        # use these to adapt the hyperparameters of the dynamics
-        adaptation_state, info_to_be_stored = adaptation_update(
-            adaptation_state, expected_value_summary_statistics
-        )
-
-        return (state, adaptation_state), info_to_be_stored
-
-    return add_all_chains_info(
-        add_splitR(step, num_chains, superchain_size), all_chains_info
-    )
-
-
-def add_splitR(step, num_chains, superchain_size):
-    def _step_with_R(state_all, xs):
-        state_all, info_to_be_stored = step(state_all, xs)
-
-        state, adaptation_state = state_all
-
-        R = splitR(state.position, num_chains, superchain_size)
-        split_bavg = jnp.average(jnp.square(R) - 1)
-        split_bmax = jnp.max(jnp.square(R) - 1)
-
-        info_to_be_stored["R_avg"] = split_bavg
-        info_to_be_stored["R_max"] = split_bmax
-
-        return (state, adaptation_state), info_to_be_stored
-
-    def _step_with_R_1(state_all, xs):
-        state_all, info_to_be_stored = step(state_all, xs)
-
-        info_to_be_stored["R_avg"] = 0.0
-        info_to_be_stored["R_max"] = 0.0
-
-        return state_all, info_to_be_stored
-
-    if superchain_size is None:
-        return step
-
-    if superchain_size == 1:
-        return _step_with_R_1
-
-    else:
-        return _step_with_R
-
-
-def add_all_chains_info(step, all_chains_info):
-    def _step(state_all, xs):
-        (state, adaptation_state), info_to_be_stored = step(state_all, xs)
-        info_to_be_stored["all_chains_info"] = vmap(all_chains_info)(state.position)
-
-        return (state, adaptation_state), info_to_be_stored
-
-    return _step if all_chains_info is not None else step
-
-
-def while_with_info(step, init, xs, length, while_cond):
-    """Same syntax and usage as jax.lax.scan, but it is run as a while loop that is terminated if not while_cond(state).
-    len(xs) determines the maximum number of iterations.
-    """
-
-    get_i = lambda tree, i: jax.tree.map(lambda arr: arr[i], tree)
-
-    info1 = step(init, get_i(xs, 0))[
-        1
-    ]  # call the step once to determine the shape of info
-    info = jax.lax.scan(lambda x, _: (x, info1), init=0, length=length)[
-        1
-    ]  # allocate the full info by repeating values
-
-    init_val = (init, info, 0, while_cond(info1, 0))
-
-    def body_fun(val):
-        x, info_old, counter, cond = val
-
-        x_new, info_new = step(x, get_i(xs, counter))
-
-        # update the full info by adding the new one
-        info_full = jax.tree.map(
-            lambda arr, val: arr.at[counter].set(val), info_old, info_new
-        )
-
-        cond = while_cond(info_new, counter)
-
-        return x_new, info_full, counter + 1, cond
-
-    def cond_fun(val):
-        _, _, counter, cond = val
-        return cond & (counter < length)
-
-    final, info, counter, _ = jax.lax.while_loop(cond_fun, body_fun, init_val)
-
-    return final, info, counter
-
-
-def run_eca(
-    rng_key,
-    initial_state,
-    kernel,
-    adaptation,
-    num_steps,
-    num_chains,
-    mesh,
-    superchain_size=None,
-    all_chains_info=None,
-    early_stop=False,
-):
-    """
-    Run ensemble chain adaptation (eca) in parallel on multiple devices.
-    -----------------------------------------------------
-    Args:
-        rng_key: random key
-        initial_state: initial state of the system
-        kernel: kernel for the dynamics
-        adaptation: adaptation object
-        num_steps: number of steps to run
-        num_chains: number of chains
-        mesh: mesh for parallelization
-        all_chains_info: function that takes the state of the system and returns some summary statistics. Will be applied and stored for all the chains at each step so it can be memory intensive.
-        early_stop: whether to stop early
-    Returns:
-        final_state: final state of the system
-        final_adaptation_state: final adaptation state
-        info_history: history of the information that was stored at each step (if early_stop is False, then this is None)
-    """
-
-    step = eca_step(
-        kernel,
-        adaptation.summary_statistics_fn,
-        adaptation.update,
-        num_chains,
-        superchain_size=superchain_size,
-        all_chains_info=all_chains_info,
-    )
-
-    def all_steps(initial_state, keys_sampling, keys_adaptation):
-        """This function operates on a single device. key is a random key for this device."""
-
-        initial_state_all = (initial_state, adaptation.initial_state)
-
-        # run sampling
-        xs = (
-            jnp.arange(num_steps),
-            keys_sampling.T,
-            keys_adaptation,
-        )  # keys for all steps that will be performed. keys_sampling.shape = (num_steps, chains_per_device), keys_adaptation.shape = (num_steps, )
-
-        if early_stop:
-            final_state_all, info_history, counter = while_with_info(
-                step, initial_state_all, xs, num_steps, adaptation.while_cond
-            )
-
-        else:
-            final_state_all, info_history = lax.scan(step, initial_state_all, xs)
-            counter = num_steps
-
-        final_state, final_adaptation_state = final_state_all
-
-        return (
-            final_state,
-            final_adaptation_state,
-            info_history,
-            counter,
-        )  # info history is composed of averages over all chains, so it is a couple of scalars
-
-    p, pscalar = PartitionSpec("chains"), PartitionSpec()
-    parallel_execute = shard_map(
-        all_steps,
-        mesh=mesh,
-        in_specs=(p, p, pscalar),
-        out_specs=(p, pscalar, pscalar, pscalar),
-    )
-
-    # produce all random keys that will be needed
-
-    key_sampling, key_adaptation = split(rng_key)
-    num_steps = jnp.array(num_steps).item()
-    keys_adaptation = split(key_adaptation, num_steps)
-    distribute_keys = lambda key, shape: device_put(
-        split(key, shape), NamedSharding(mesh, p)
-    )  # random keys, distributed across devices
-    keys_sampling = distribute_keys(key_sampling, (num_chains, num_steps))
-
-    # run sampling in parallel
-    final_state, final_adaptation_state, info_history, counter = parallel_execute(
-        initial_state, keys_sampling, keys_adaptation
-    )
-
-    # info_history has a static size, determined by num_steps, but if early_stop = True, the values after the while condition has been violated are nonsense. Remove them:
-    info_history = jax.tree.map(lambda arr: arr[: int(counter)], info_history)
-
-    return final_state, final_adaptation_state, info_history
-
-
-def ensemble_execute_fn(
-    func,
-    rng_key,
-    num_chains,
-    mesh,
-    x=None,
-    args=None,
-    summary_statistics_fn=lambda y: 0.0,
-    superchain_size=None,
-):
-    """Given a sequential function
-     func(rng_key, x, args) = y,
-    evaluate it with an ensemble and also compute some summary statistics E[theta(y)], where expectation is taken over ensemble.
-    Args:
-         x: array distributed over all decvices
-         args: additional arguments for func, not distributed.
-         summary_statistics_fn: operates on a chain and returns some summary statistics.
-         rng_key: a single random key, which will then be split, such that chain will get a different random key.
-
-    Returns:
-         y: array distributed over all decvices. Need not be of the same shape as x.
-         Etheta: expected values of the summary statistics
-    """
-    p, pscalar = PartitionSpec("chains"), PartitionSpec()
-
-    if x is None:
-        X = device_put(jnp.zeros(num_chains), NamedSharding(mesh, p))
-    else:
-        X = x
-
-    adaptation_update = lambda _, Etheta: (Etheta, None)
-
-    _F = eca_step(
-        func,
-        lambda y, info, key: summary_statistics_fn(y),
-        adaptation_update,
-        num_chains,
-    )
-
-    def F(x, keys):
-        """This function operates on a single device. key is a random key for this device."""
-        y, summary_statistics = _F((x, args), (None, keys, None))[0]
-        return y, summary_statistics
-
-    parallel_execute = shard_map(F, mesh=mesh, in_specs=(p, p), out_specs=(p, pscalar))
-
-    if superchain_size == 1:
-        _keys = split(rng_key, num_chains)
-
-    else:
-        _keys = jnp.repeat(
-            split(rng_key, num_chains // superchain_size), superchain_size
-        )
-
-    keys = device_put(
-        _keys, NamedSharding(mesh, p)
-    )  # random keys, distributed across devices
-
-    # apply F in parallel
-    return parallel_execute(X, keys)
-
-
 def thin_algorithm(
     sampling_algorithm: SamplingAlgorithm,
     thinning: int = 1,
@@ -675,144 +385,6 @@ def thin_algorithm(
         return state, info_transform(info)
 
     return SamplingAlgorithm(sampling_algorithm.init, step_fn)
-
-
-def _gpdfit(exceedances: Array) -> tuple[Array, Array]:
-    """Empirical Bayes GPD fit via Zhang & Stephens (2009).
-
-    Fits a Generalised Pareto Distribution to ``exceedances`` (a 1-D array of
-    non-negative values in *ascending* order) using the Bayesian model-averaging
-    estimator.  All operations on fixed-shape arrays so the function is
-    JIT-compatible (``exceedances.shape[0]`` must be static).
-    """
-    n = exceedances.shape[0]  # static at trace time
-    prior_bs, prior_k = 3, 10
-    m_est = 30 + int(n**0.5)  # static
-
-    # Guard against the degenerate case where all exceedances are zero
-    # (e.g. all importance weights equal).  In that case k=0, sigma=0
-    # and the caller will leave the tail unchanged.
-    tiny = jnp.finfo(exceedances.dtype).tiny
-    is_degenerate = exceedances[-1] < tiny
-    safe_exc = jnp.where(is_degenerate, jnp.ones_like(exceedances), exceedances)
-
-    # Grid of candidate rate parameters b.
-    b = 1.0 - jnp.sqrt(m_est / (jnp.arange(1, m_est + 1) - 0.5))
-    b = b / (prior_bs * safe_exc[int(n / 4 + 0.5) - 1])
-    b = b + 1.0 / safe_exc[-1]
-
-    # k estimate for each b (mean log1p).
-    k_ary = jnp.mean(jnp.log1p(-b[:, None] * safe_exc[None, :]), axis=1)
-
-    # Profile log-likelihood weights.
-    len_scale = n * (jnp.log(-b / k_ary) - k_ary - 1.0)
-    w = jnp.exp(len_scale - jax.nn.logsumexp(len_scale))
-
-    # Posterior mean of b, then derive k and sigma.
-    b_post = jnp.sum(b * w)
-    k = jnp.mean(jnp.log1p(-b_post * safe_exc))
-    sigma = -k / b_post
-    # Weakly informative prior shrinks k toward 0.5.
-    k = (n * k + prior_k * 0.5) / (n + prior_k)
-
-    k = jnp.where(is_degenerate, jnp.zeros(()), k)
-    sigma = jnp.where(is_degenerate, jnp.zeros(()), sigma)
-    return k, sigma
-
-
-def _gpinv(p: Array, k: Array, sigma: Array) -> Array:
-    """GPD quantile function (inverse CDF)."""
-    return jnp.where(
-        jnp.abs(k) < 1e-10,
-        -sigma * jnp.log1p(-p),
-        sigma * jnp.expm1(-k * jnp.log1p(-p)) / k,
-    )
-
-
-def psis_weights(log_ratios: Array, r_eff: float = 1.0) -> tuple[Array, Array]:
-    """Pareto Smoothed Importance Sampling (PSIS) log weights.
-
-    Implements the PSIS smoothing step from :cite:p:`vehtari2017practical`:
-    the ``M`` largest importance ratios (in ratio space) are replaced by sorted
-    Generalised Pareto quantiles fitted by the empirical Bayes estimator of
-    Zhang & Stephens (2009), then all weights are normalised.
-
-    This is a pure-JAX, JIT-compatible implementation faithful to Algorithm 1
-    of Vehtari, Gelman & Gabry (2017).
-
-    Parameters
-    ----------
-    log_ratios
-        Log importance ratios ``log p(θ) − log q(θ)``, shape ``(n,)``.
-        Need not be normalised.
-    r_eff
-        Relative effective sample size of the proposal, ``S_eff / n``.
-        Use the default of ``1.0`` for i.i.d. draws (e.g. Pathfinder);
-        set to the actual ESS ratio for correlated MCMC chains.  Values
-        below 1 increase the tail size ``M`` to compensate for correlation.
-
-    Returns
-    -------
-    log_weights
-        Normalised log importance weights, shape ``(n,)``.
-        ``jnp.exp(log_weights).sum() == 1`` up to floating-point precision.
-    pareto_k
-        Pareto shape parameter estimate (scalar ``Array``).  Values below 0.5
-        indicate reliable estimates; 0.5–0.7 are moderate; above 0.7 may give
-        unreliable estimates.  ``jnp.inf`` means the tail was too small to fit
-        (fewer than 5 samples).
-
-    Notes
-    -----
-    Tail size: ``M = min(floor(3*sqrt(n/r_eff)), n//5)``, matching the paper.
-    The GPD is only applied when ``k >= 1/3``; lighter tails are left
-    unsmoothed (only normalised).  Fitting uses empirical Bayes in
-    importance-ratio space, the same approach as ArviZ.
-    """
-    n = log_ratios.shape[0]
-    M = min(max(int(3.0 * (n / r_eff) ** 0.5), 5), n // 5)
-
-    if M < 5:
-        # Too few tail samples for a reliable GPD fit; return normalised
-        # weights and k=inf to signal this to the caller.
-        log_w = log_ratios - jax.nn.logsumexp(log_ratios)
-        return log_w, jnp.asarray(jnp.inf)
-
-    # Stabilise numerically.
-    lw = log_ratios - log_ratios.max()
-
-    # Sort ascending so that the M largest are at positions [n-M:].
-    sorted_idx = jnp.argsort(lw)
-    lw_sorted = lw[sorted_idx]
-
-    # Threshold: largest value below the tail (in log and ratio space).
-    threshold_log = lw_sorted[n - M - 1]
-    threshold_ratio = jnp.exp(threshold_log)
-
-    # Work in ratio (non-log) space for GPD fitting, as in the original paper.
-    tail_ratio = jnp.exp(lw_sorted[n - M :])  # ascending, shape (M,)
-    exceedances = tail_ratio - threshold_ratio  # >= 0, ascending
-
-    k, sigma = _gpdfit(exceedances)
-
-    # Uniform quantile positions within the tail: (0.5/M, 1.5/M, ..., (M-0.5)/M).
-    p = (jnp.arange(M) + 0.5) / M  # ascending
-    smoothed = threshold_ratio + _gpinv(p, k, sigma)
-    # Cap smoothed values at the observed tail maximum.
-    smoothed = jnp.minimum(smoothed, tail_ratio[-1])
-
-    # Only replace the tail when k >= 1/3 (paper Algorithm 1, step 7).
-    # For lighter tails the raw order statistics are already reliable.
-    lw_smooth = jnp.where(
-        k >= 1.0 / 3.0,
-        lw_sorted.at[n - M :].set(jnp.log(smoothed)),
-        lw_sorted,
-    )
-
-    # Restore original ordering and normalise.
-    lw_orig = jnp.zeros_like(lw_smooth).at[sorted_idx].set(lw_smooth)
-    log_w = lw_orig - jax.nn.logsumexp(lw_orig)
-    return log_w, k
 
 
 def thin_kernel(
